+++ capcom/capcom.hpp
@@ -33,7 +33,7 @@ namespace capcom
 		uintptr_t get_export(uintptr_t base, uint16_t ordinal);
 		uintptr_t get_export(uintptr_t base, const char* name);
 		uintptr_t allocate_pool(size_t size, kernel::POOL_TYPE pool_type, const bool page_align, size_t* out_size = nullptr);
-		uintptr_t allocate_pool(std::size_t size, uint16_t pooltag, kernel::POOL_TYPE = kernel::NonPagedPool, bool page_align = false, size_t* out_size = nullptr);
+		uintptr_t allocate_pool(std::size_t size, uint32_t pooltag, kernel::POOL_TYPE = kernel::NonPagedPool, bool page_align = false, size_t* out_size = nullptr);
 		template <typename T>
 		T get_system_routine(const std::wstring& name)
 		{
